1. Added Device model, manufacturer, Size(partial) 2. Apis for getting values for each dimension 3. Concurrent Hbase calls 4. Refactored enums 5. Added keywords for few enums 6. Added controllers

+++ src/com/tapjoy/reach/counts/CountsModel.java
@@ -1,37 +1,37 @@
 package com.tapjoy.reach.counts;
 
 public class CountsModel {
-	private int udids_count;
-	private int impressions_count;
-	private int unique_viewers_count;
+	private long udids_count;
+	private long impressions_count;
+	private long unique_viewers_count;
 	
-	public CountsModel(int udidsCount, int offerwallImpCount, int uvCount){
+	public CountsModel(long udidsCount, long offerwallImpCount, long uvCount){
 		this.udids_count = udidsCount;
 		this.impressions_count = offerwallImpCount;
 		this.unique_viewers_count = uvCount;
 	}
 
-	public int getUdids_count() {
+	public long getUdids_count() {
 		return udids_count;
 	}
 
-	public void setUdids_count(int udids_count) {
+	public void setUdids_count(long udids_count) {
 		this.udids_count = udids_count;
 	}
 
-	public int getImpressions_count() {
+	public long getImpressions_count() {
 		return impressions_count;
 	}
 
-	public void setImpressions_count(int impressions_count) {
+	public void setImpressions_count(long impressions_count) {
 		this.impressions_count = impressions_count;
 	}
 	
-	public int getUnique_viewers_count() {
+	public long getUnique_viewers_count() {
 		return unique_viewers_count;
 	}
 
-	public void setUnique_viewers_count(int unique_viewers_count) {
+	public void setUnique_viewers_count(long unique_viewers_count) {
 		this.unique_viewers_count = unique_viewers_count;
 	}
